Pipeline releases page - expand first release
On the pipeline releases page, each release shows a toggle expand with details and buttons. This isn't super obvious from just looking at it. To help, this change makes the top release expanded on page load. This is usually the one that people will be interested in anyway, and also hints at the functionality for the others.