Update university-of-miami.mdx
we are changing the node to the Sylvester Cancer Center, it is part of University of Miami but Sylvester is our main place of work at the UM's Miller School of Medicine. We will soon comfirm the venue, we are trying to book a nice room with good wifi, as our meeting room may be not ideal. thanks!